feat: separate `use-wallet` functions to a package
