+++ packages/src/@terra-dev/use-wallet/useWallet.ts
@@ -1,8 +1,8 @@
 import { NetworkInfo } from '@terra-dev/wallet-types';
 import { CreateTxOptions } from '@terra-money/terra.js';
 import { Consumer, Context, createContext, useContext } from 'react';
-import { TxResult } from '../tx';
-import { ConnectType, WalletInfo, WalletStatus } from '../types';
+import { TxResult } from './tx';
+import { ConnectType, WalletInfo, WalletStatus } from './types';
 
 export interface Wallet {
   /**
