+++ packages/src/@terra-money/wallet-provider/index.ts
@@ -1,6 +1,5 @@
-export * from './types';
 export * from './controller';
-export * from './tx';
 export * from './react';
 
+export * from '@terra-dev/use-wallet';
 export * from '@terra-dev/wallet-types';
