+++ packages/src/@terra-money/wallet-provider/react/index.ts
@@ -1,6 +1,7 @@
+export * from '@terra-dev/use-wallet/useWallet';
+export * from '@terra-dev/use-wallet/useConnectedWallet';
+
 export * from './ExtensionNetworkOnlyWalletProvider';
-export * from './useConnectedWallet';
-export * from './useWallet';
 export * from './WalletProvider';
 export * from './useInstallChromeExtension';
 export * from './StaticWalletProvider';
